logs collection

+++ sand2/log.js
@@ -1,19 +1,38 @@
 // log.js
 //  append log strings to an existing file (log.txt)
+//
+//
+//  "logs" collection thoughts:
+//  {
+//      contest_id = NNN
+//      user_id = NNN
+//      logNumber = NNN  (incrementing number per log entry)
+//      msg = "log message"
+//  }
+var mdbURL = process.env.MONGOLAB_URI;   // URL to THE database
+var isLocal = (mdbURL? false : true);
 
 var fs = require('fs');
 var path = require('path');
+var logdb = require('logsb');
 var logfile = path.join(__dirname, "log.txt");
 var prefix = "";
+var logN = 1;
 
 var log = function(msg) {
-    var str = "" + (new Date().toString()) + ": "+prefix+": " + msg + "\n";
-    if (fs.appendFileSync) {
-        fs.appendFileSync(logfile, str);
+    if (isLocal) {
+        var str = "" + (new Date().toString()) + ": "+prefix+": " + msg + "\n";
+        if (fs.appendFileSync) {
+            fs.appendFileSync(logfile, str);
+        }
+    } else {
+        // @TODO: ONLY record if:  CHILD[P1] or CHILD[P2]
+        var doc = {};
+        doc.date = "" + (new Date().toString());
+        doc.msg = msg;
+        doc.logNumber = logN++;
+        logdb.postLogs(doc);
     }
-//    var id = fs.openSync(logfile, 'a');
-//    fs.writeSync( id, str, 0, str.length, null);
-//    fs.closeSync(id);
 };
 
 var setPrefix = function(str) {
@@ -22,48 +41,53 @@ var setPrefix = function(str) {
 };
 
 var resetLogFile = function() {
-    try {
-        fs.unlinkSync(logfile);
-    } catch (e) {
-        // ignore error (assume file-not-found)
+    if (isLocal) {
+        try {
+            fs.unlinkSync(logfile);
+        } catch (e) {
+            // ignore error (assume file-not-found)
+        }
+        log("- - - - - - - NEW CONTEST STARTING - - - - - -");
     }
-    log("- - - - - - - NEW CONTEST STARTING - - - - - -");
 };
 
 // get all log messages (by player#) -- return array of log messages
 // return:  logs[0] = ["Player 1 log message 1", "log message 2" ... ]
 //          logs[1] = ["Player 2 log message 1", "log message 2" ... ]
 var getLogMessages = function(fnc) {
-    var logMessages = [];
-    logMessages[0] = [];
-    logMessages[1] = [];
-    // Fri Sep 28 2012 20:51:56 GMT-0600 (MDT): CHILD[P2]: USER_LOG:DANB calling submitTurn with pick=r
-    // ----------------------------------------"CHILD[P" <number> "]: USER_LOG:"
-    // Fri Sep 28 2012 20:51:56 GMT-0600 (MDT): DANB calling submitTurn with pick=r
+    if (isLocal) {
+        var logMessages = [];
+        logMessages[0] = [];
+        logMessages[1] = [];
+        // Fri Sep 28 2012 20:51:56 GMT-0600 (MDT): CHILD[P2]: USER_LOG:DANB calling submitTurn with pick=r
+        // ----------------------------------------"CHILD[P" <number> "]: USER_LOG:"
+        // Fri Sep 28 2012 20:51:56 GMT-0600 (MDT): DANB calling submitTurn with pick=r
 
-    var input = fs.createReadStream(logfile);
-    readLines(input, function(oneLine, isDone) {
-        if (isDone) {
-            fnc(logMessages);
-        } else {
-            var str1 = "CHILD[P";
-            var str2 = "]: USER_LOG:";
-            var idx1 = oneLine.indexOf(str1);
-            var idx2 = oneLine.indexOf(str2);
-            if (idx1>0 && idx2>0) {
-                try {
-                    var pn = parseInt(oneLine.substr(idx1+str1.length, 1));
-                    var msg = oneLine.substr(0, idx1) + oneLine.substr(idx2+str2.length);
-                    if (pn >= 1 && pn <= 2) {
-                        logMessages[pn-1] = msg;
+        var input = fs.createReadStream(logfile);
+        readLines(input, function(oneLine, isDone) {
+            if (isDone) {
+                fnc(logMessages);
+            } else {
+                var str1 = "CHILD[P";
+                var str2 = "]: USER_LOG:";
+                var idx1 = oneLine.indexOf(str1);
+                var idx2 = oneLine.indexOf(str2);
+                if (idx1>0 && idx2>0) {
+                    try {
+                        var pn = parseInt(oneLine.substr(idx1+str1.length, 1));
+                        var msg = oneLine.substr(0, idx1) + oneLine.substr(idx2+str2.length);
+                        if (pn >= 1 && pn <= 2) {
+                            logMessages[pn-1] = msg;
+                        }
+                    } catch (e) {
+                        // ignore error ... bad line ... ignore it
                     }
-                } catch (e) {
-                    // ignore error ... bad line ... ignore it
                 }
             }
-        }
-    });
-
+        });
+    } else {
+        fnc();
+    }
 };
 
 // read an entire file, one line at a time
